% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.r
\name{fbind}
\alias{fbind}
\title{Fbind() Une dos factores}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{factor}

\item{b}{factor}
}
\value{
factor
}
\description{
Crea un factor a partir de 2 factores nuevos
haciendo la union de los 2 factores unidos
}
\examples{
fbind(iris$Species[c(1, 51, 101)], PlantGrowth$group[c(1, 11, 21)])
}
