% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcount.R
\name{fcount}
\alias{fcount}
\title{Fcount() Haciendo una frecuencia ordenada de un factor}
\usage{
fcount(x)
}
\arguments{
\item{x}{factor}
}
\value{
A tibble
}
\description{
Fcount() Haciendo una frecuencia ordenada de un factor
}
\examples{
fcount(iris$Species)
}
